fix(security): resolve validation/notifications circular import
Moving validatePublicServiceUrl into validation.ts created a circular
import (validation <-> notifications). The bundler handled it at
typecheck but at runtime the import cycle left isPrivateIp undefined
for some callers, which broke the dev server boot and every e2e test.

Consolidate the URL safety helpers (isPrivateIp, isPrivateUrl,
normalizeIp, isCloudMetadata, validatePublicServiceUrl) in
validation.ts. notifications.ts now re-exports isPrivateIp and
isPrivateUrl so the auth/http/settings callers keep working.
